fs-1031; Part one fixes
