+++ cmd/inventory.go
@@ -5,20 +5,22 @@ import (
 
 	"github.com/equinix-labs/otel-init-go/otelinit"
 	"github.com/metal-toolbox/fleet-scheduler/internal/app"
+	"github.com/metal-toolbox/fleet-scheduler/internal/client"
+	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"golang.org/x/net/context"
 )
 
-var cmdCollect = &cobra.Command{
-	Use:   "collect",
-	Short: "collect em",
+var cmdInventory = &cobra.Command{
+	Use:   "inventory",
+	Short: "gather all servers and create invetory for them",
 	Run: func(cmd *cobra.Command, args []string) {
 		collect(cmd.Context())
 	},
 }
 
 func init() {
-	RootCmd.AddCommand(cmdCollect)
+	rootCmd.AddCommand(cmdInventory)
 }
 
 func collect(ctx context.Context) {
@@ -34,8 +36,10 @@ func collect(ctx context.Context) {
 		return
 	}
 
-	client, err := app.NewClient()
-		if err != nil {
+	loggerEntry := app.Logger.WithFields(logrus.Fields{"component": "store.serverservice"})
+	loggerEntry.Level = app.Logger.Level
+	client, err := client.New(app.Ctx, app.Cfg, loggerEntry)
+	if err != nil {
 		log.Fatal(err)
 		return
 	}
