+++ config.yaml
@@ -1,6 +1,6 @@
 log_level: debug
 store_kind: fleetscheduler
-serverservice:
+fleetdb:
   endpoint: http://192.168.1.191:8000
   disable_oauth: true
   facility_code: fr9
