+++ internal/client/conditionorc.go
@@ -4,27 +4,27 @@ import (
 	"encoding/json"
 
 	"github.com/google/uuid"
-	coapi "github.com/metal-toolbox/conditionorc/pkg/api/v1/types"
-	rivetsCondition "github.com/metal-toolbox/rivets/condition"
+	conditionOrcapi "github.com/metal-toolbox/conditionorc/pkg/api/v1/types"
+	conditionrivets "github.com/metal-toolbox/rivets/condition"
 )
 
-func (c *Client) CreateCondition(serverUUID uuid.UUID) error {
-	params, err := json.Marshal(rivetsCondition.InventoryTaskParameters{
+func (c *Client) CreateConditionInventory(serverUUID uuid.UUID) error {
+	params, err := json.Marshal(conditionrivets.InventoryTaskParameters{
 		AssetID:               serverUUID,
 		CollectBiosCfg:        false,
 		CollectFirwmareStatus: false,
-		Method:                rivetsCondition.OutofbandInventory,
+		Method:                conditionrivets.OutofbandInventory,
 	})
 	if err != nil {
 		return err
 	}
 
-	conditionCreate := coapi.ConditionCreate {
+	conditionCreate := conditionOrcapi.ConditionCreate {
 		Exclusive: false,
 		Parameters: params,
 	}
 
-	_, err = c.coClient.ServerConditionCreate(c.ctx, serverUUID, rivetsCondition.Inventory, conditionCreate)
+	_, err = c.coClient.ServerConditionCreate(c.ctx, serverUUID, conditionrivets.Inventory, conditionCreate)
 	if err != nil {
 		return err
 	}
